backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
selenium==4.15.0
webdriver-manager==4.0.1
requests==2.31.0